fix(swizdb): add check to updater for function availability

+++ scripts/update/00-checkreqs.sh
@@ -2,6 +2,9 @@
 # Checks if required functions are loaded before continuing with the update scripts.
 # Function should be loaded on the next run (by `box`) without further need for interaction.
 
+# DO NOT SOURCE GLOBALS HERE, THAT DEFEATS THE WHOLE PURPOSE OF THIS CHECK
+# SOURCING GLOBALS HERE WILL MAKE THESE TESTS PASS BUT THE FUNCTIONS WON'T BE AVAILABLE IN THE SHELL ABOVE
+
 if ! command -v apt_install > /dev/null 2>&1; then
     echo
     echo "Due to internal restructuring please restart \`box update\`. You should only have to do this once."
@@ -22,6 +25,14 @@ if [[ -z $log ]]; then
     echo
     echo "Due to internal restructuring please restart \`box update\`. You should only have to do this once."
     echo "Reason: log not set"
+    kill -13 $(ps --pid $$ -oppid=)
+    exit 1
+fi
+
+if ! swizdb list; then
+    echo
+    echo "Due to internal restructuring please restart \`box update\`. You should only have to do this once."
+    echo "Reason: swizdb licommand unavailable"
     kill -13 $(ps --pid $$ -oppid=)
     exit 1
 fi
